Make KeyCardPdf a separate class. Use wallet response params to generate PDF
