+++ src/lib/resources/bitbay.png
